Update on the assignment
There are still problems even seeing the site. I can only see the pages if I'm in the dashboard and on the blog page. I then select, "see page" and the site comes up.

I checked with GoDaddy and they said since they could see my info in other themes they couldn't help me with the code in mine. 

I will try to keep going with this but I'm not sure what else to do.
David